backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
pdfplumber>=0.11.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0
ddgs>=5.2.2
sentence-transformers>=2.2.0
chromadb>=0.4.0
pydantic>=2.0
python-dotenv>=1.0.0
plotly>=5.18.0
